feat: add check for products with category to DeleteProductCategory action

+++ src/DevBook.API/Features/BookStore/Products/ProductCategories/DeleteProductCategory.cs
@@ -8,7 +8,14 @@ internal class DeleteProductCategoryCommandHandler(DevBookDbContext dbContext) :
 	{
 		var productCategory = await dbContext.ProductCategories.FindAsync([command.Id], cancellationToken);
 
-		// Check for books with this category and throw DevBookError when there are books in this category => cannot delete category with products
+		var productWithCategory = await dbContext.Products.FirstOrDefaultAsync(x => x.ProductCategoryIds.Contains(command.Id));
+		if (productWithCategory is not null)
+		{
+			throw new DevBookValidationException(
+				nameof(command.Id),
+				$"ProductCategory with id '{command.Id}' is used on some Products and cannot be deleted. To delete ProductCategory, you need to remove it from Products first.");
+		}
+
 		if (productCategory is not null)
 		{
 			dbContext.ProductCategories.Remove(productCategory);
